* master:
   update version info
   NetBox.vcxproj: remove gc-7.2alpha6
   revert "RemoteFiles.cpp: try to fix AV"
  Revert "RemoteFiles.cpp: temporarily remove delete object"
   update version info
   SessionInfo: minor
   ConvertTimestampToUTC: fix
  fix warnings
   fix C4239

Conflicts:
	src/NetBox/NetBox.vcxproj
	src/NetBox/WinSCPDialogs.cpp

+++ src/core/CoreMain.cpp
@@ -46,7 +46,7 @@ bool IsAuthenticationPrompt(TPromptKind Kind)
 void CoreInitialize()
 {
   // Randomize();
-  srand (time(NULL));
+  srand(static_cast<size_t>(time(NULL)));
   CryptographyInitialize();
 
   // configuration needs to be created and loaded before putty is initialized,
