+++ src/core/Cryptography.cpp
@@ -550,8 +550,8 @@ bool UnscramblePassword(std::wstring & Password)
   bool Result = false;
   if (strlen(S) >= 3)
   {
-    int Len = (S[0] - '0') + 10 * (S[1] - '0') + 100 * (S[2] - '0');
-    int Total = (((Len + 3) / 17) * 17 + 17);
+    size_t Len = (S[0] - '0') + 10 * (S[1] - '0') + 100 * (S[2] - '0');
+    size_t Total = (((Len + 3) / 17) * 17 + 17);
     if ((Len >= 0) && (Total == Password.size()) && (Total - (S - (char *)Password.c_str()) - 3 == Len))
     {
       Password.erase(Password.size() - Len, 1);
